+++ src/main/java/edu/ie3/datamodel/io/naming/DefaultDirectoryHierarchy.java
@@ -255,6 +255,7 @@ public class DefaultDirectoryHierarchy implements FileHierarchy {
         Constants.INPUT_SUB_TREE.resolve("participants"),
         true,
         Stream.of(
+                AcInput.class,
                 BmInput.class,
                 ChpInput.class,
                 EvInput.class,
@@ -270,6 +271,7 @@ public class DefaultDirectoryHierarchy implements FileHierarchy {
         Constants.RESULT_SUB_TREE.resolve("participants"),
         false,
         Stream.of(
+                AcResult.class,
                 BmResult.class,
                 ChpResult.class,
                 EvResult.class,